requirements.txt
nest-asyncio==1.5.6
recnetlogin==0.0.5
recnetpy==0.1.38
python-dotenv==0.21.0 
pypresence==4.2.1
httpx==0.23.0